requirements.txt
tensorflow>=2.12.0
gradio>=3.41.0
pandas>=1.5.0
numpy>=1.23.0
Pillow>=9.0.0
scikit-learn>=1.0.0
